% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint-dirs.R
\name{lint_dir}
\alias{lint_dir}
\title{Lint a directory}
\usage{
lint_dir(path = ".", recursive = FALSE, relative_path = TRUE, ...)
}
\arguments{
\item{path}{The path to the directory to lint.}

\item{recursive}{Recursively search directories}

\item{relative_path}{If \code{TRUE}, file paths are printed using
their path relative to the package base directory.
If \code{FALSE}, use the full absolute path.}

\item{...}{Arguments passed to \code{link[lintr]{lint}}}
}
\description{
Lint a directory
}
