% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lint-dirs.R
\name{lint_root_dir}
\alias{lint_root_dir}
\alias{lint_rstudio_project}
\alias{lint_r_package}
\title{Lint R projects or packages}
\usage{
lint_root_dir(criterion, ..., path = ".")

lint_rstudio_project(criterion, ..., path = ".")

lint_r_package(criterion, ..., path = ".")
}
\arguments{
\item{criterion}{A criterion, will be coerced using \code{\link[rprojroot]{as.root_criterion}()}.}

\item{...}{Arguments passed to \code{\link{lint_dir}}}

\item{path}{Directory to start searching for the root directory of the
project or}
}
\description{
Lint R projects or packages
}
