% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{linters}
\alias{linters}
\alias{function_left_parentheses_linter}
\title{Miscellaneous Linters}
\usage{
function_left_parentheses_linter(source_file)
}
\arguments{
\item{source_file}{returned by \code{\link[lintr]{get_source_expressions}}}
}
\value{
An \code{lints} object, which is a list of \code{lint} objects.
}
\description{
Additional linters
}
